🚑️ Add temporary fix for some anomalies
For example, in some cases, the name of the school and the name of the
group (e.g., "BUSINESS STUDIES") are in the same line, which the parser
cannot parse.

The number of instances like this is quite small: probably within a
margin of 2.5% (based on the fact that the number of schools parsed
from for `ssc-2019` and `ssc-2020` are 16,985 and 17,409 respectively).
So, perhaps these cases can be handled manually. Though, a proper fix to
this issue would be attempted later.